Move voting, saving, and score calculation into VotingRound
- Most helper functions for VotingRound are now moved into VotingRound itself.
- File management and digit counting is moved into helper file.
- currentOptionItems was a test-specific function. It is now moved into its corresponding test file.
