+++ src/test/test_load_and_save_file.cpp
@@ -2,6 +2,7 @@
 #include <fstream>
 
 #include "functions.h"
+#include "helpers.h"
 #include "testing.h"
 
 namespace
